+++ Omise/Client.cs
@@ -62,6 +62,7 @@ namespace Omise
 
         public ChargeSpecificResource Charge(string chargeId) => new ChargeSpecificResource(requester, chargeId);
         public CustomerSpecificResource Customer(string customerId) => new CustomerSpecificResource(requester, customerId);
+        public RecipientSpecificResource Recipient(string recipientId) => new RecipientSpecificResource(requester, recipientId);
         public ScheduleSpecificResource Schedule(string scheduleId) => new ScheduleSpecificResource(requester, scheduleId);
     }
 }