+++ Omise/Resources/TransferScheduleResource.cs
@@ -1,0 +1,13 @@
+﻿using Omise.Models;
+
+namespace Omise.Resources
+{
+    public class TransferScheduleResource : BaseResource<Schedule>,
+    IListable<Schedule>
+    {
+        public TransferScheduleResource(IRequester requester)
+            : base(requester, Endpoint.Api, "/transfers/schedules")
+        {
+        }
+    }
+}