+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # CHANGE LOG
 
+# v2.5.4
+
+* **NEW:** Transfer schedule APIs.
+* **FIXED:** There is no `currency` field on TransferScheduling, it has been removed.
+* **FIXED:** `Charges.Schedules` resource previously returns wrong type `Charge` (should be
+  `Schedule`)
+
 # v2.5.3
 
 * **FIXED:** `installment_terms` are sent as `0` when not set. It is now changed to a